refactor: colocate rollback achievement progress tests
